Refactor seed data (join all cars in one file, add all manufacturers before)
