+++ prisma/seed/data/car_manufacturers.ts
@@ -1,0 +1,61 @@
+import { Prisma } from '@prisma/client'
+
+export const manufacturersData: Prisma.CarManufacturerCreateInput[] = [
+  {
+    name: 'Aston Martin'
+  },
+  {
+    name: 'Audi'
+  },
+  {
+    name: 'BMW'
+  },
+  {
+    name: 'Bugatti'
+  },
+  {
+    name: 'Chevrolet'
+  },
+  {
+    name: 'Ford'
+  },
+  {
+    name: 'Honda'
+  },
+  {
+    name: 'Jaguar'
+  },
+  {
+    name: 'Lamborghini'
+  },
+  {
+    name: 'Lexus'
+  },
+  {
+    name: 'Maserati'
+  },
+  {
+    name: 'Mazda'
+  },
+  {
+    name: 'Mercedes-Benz'
+  },
+  {
+    name: 'Mitsubishi'
+  },
+  {
+    name: 'Nissan'
+  },
+  {
+    name: 'Porsche'
+  },
+  {
+    name: 'Tesla'
+  },
+  {
+    name: 'Toyota'
+  },
+  {
+    name: 'Volkswagen'
+  }
+]
