--- prisma/seed/data/cars/index.ts
@@ -1,19 +1,0 @@
-export * from './aston_martin'
-export * from './audi'
-export * from './bmw'
-export * from './bugatti'
-export * from './chevrolet'
-export * from './ford'
-export * from './honda'
-export * from './jaguar'
-export * from './lamborghini'
-export * from './lexus'
-export * from './maserati'
-export * from './mazda'
-export * from './mercedes-benz'
-export * from './mitsubishi'
-export * from './nissan'
-export * from './porsche'
-export * from './tesla'
-export * from './toyota'
-export * from './volkswagen'
